feat(food): add base unit and quantity nutrition calculation

+++ lib/core/models/meal_data.dart
@@ -19,7 +19,8 @@ class MealData {
 
   bool get isWaterEntry => waterMl != null;
 
-  int get calories => items.fold(0, (sum, item) => sum + item.totalCalories);
+  double get calories =>
+      items.fold(0.0, (sum, item) => sum + item.totalCalories);
 
   double get protein => items.fold(0.0, (sum, item) => sum + item.totalProtein);
